requirements.txt
environs
asn1crypto==0.24.0
astroid==2.2.5
backports.csv==1.0.7
bcrypt==3.1.7
blinker==1.4
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
colorama==0.4.1
cryptography==2.6.1
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.2
django-background-tasks==1.2.0
django-compat==1.0.15
django-cors-headers==3.0.2
django-filter==2.2.0
django-filters==0.2.1
django-import-export==1.2.0
django-oauth-toolkit==1.2.0
django-rest-framework==0.1.0
django-rest-knox==4.0.1
django-simple-history==2.7.3
django-storages==1.7.1
djangorestframework==3.9.4
et-xmlfile==1.0.1
idna==2.8
isort==4.3.18
jdcal==1.4.1
lazy-object-proxy==1.4.0
mccabe==0.6.1
oauthlib==3.0.1
odfpy==1.4.0
openpyxl==2.6.3
paramiko==2.6.0
Pillow==6.0.0
pycparser==2.19
PyJWT==1.7.1
pylint==2.3.1
PyNaCl==1.3.0
pyOpenSSL==19.0.0
PySocks==1.6.8
python-decouple==3.1
python3-openid==3.1.0
pytz==2019.1
PyYAML==5.1.2
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.1.0
sqlparse==0.3.0
tablib==0.13.0
typed-ast==1.3.4
urllib3==1.24.3
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.11.1
xlrd==1.2.0
xlwt==1.3.0
